Stuck method, first turn and promotion for pawn
All pieces are now present on the board. 
Has Aydan's stuck method in the piece class and graphics panel.
Has Aydan's first turn functions for a pawn in the pawn class and graphics panel. 
The move method in the graphics panel has code to promote a pawn to a queen.