Pawn promotion selection, game over, cleanup!
A panel shows up allowing user to select which piece their pawn will be promoted to. Pieces that are shown are objects of the PieceIcon class.
Game ends when a king is taken (the console says "___ wins!" and the user can't click any pieces). Josh is working on checkmates.
The code for the paintComponent method in the GraphicsPanel class has been split up into sub-methods.