Reorganized and added console "messages"
I organized a few things to be more logical and created a String variable called message, whose value is drawn onto the console. For example, it can say "Invalid move" or "Stuck, try again", or "Promote your pawn". Hopefully we can get it to print specifics of why certain moves are invalid.